chore: change job to be a descriminator type
