+++ application/backend/src/repositories/mappers/job_mapper.py
@@ -1,6 +1,10 @@
+from pydantic import TypeAdapter
+
 from db.schema import JobDB
 from repositories.mappers.base_mapper_interface import IBaseMapper
 from schemas import Job
+
+JOB_ADAPTER = TypeAdapter(Job)
 
 
 class JobMapper(IBaseMapper):
@@ -10,4 +14,4 @@ class JobMapper(IBaseMapper):
 
     @staticmethod
     def from_schema(model: JobDB) -> Job:
-        return Job.model_validate(model, from_attributes=True)
+        return JOB_ADAPTER.validate_python(model, from_attributes=True)
