+++ application/backend/src/schemas/__init__.py
@@ -1,7 +1,8 @@
+from .base_job import JobStatus, JobType
 from .calibration import CalibrationConfig
 from .camera import Camera, CameraProfile
 from .dataset import Dataset, Episode, EpisodeInfo, EpisodeVideo, LeRobotDatasetInfo, Snapshot
-from .job import Job
+from .job import Job, TrainJob
 from .model import Model
 from .project import Project
 from .robot import LeRobotConfig, NetworkIpRobotConfig, Robot, SerialPortInfo
@@ -15,6 +16,8 @@ __all__ = [
     "EpisodeInfo",
     "EpisodeVideo",
     "Job",
+    "JobStatus",
+    "JobType",
     "LeRobotConfig",
     "LeRobotDatasetInfo",
     "Model",
@@ -23,4 +26,5 @@ __all__ = [
     "Robot",
     "SerialPortInfo",
     "Snapshot",
+    "TrainJob",
 ]
