+++ application/backend/src/workers/training_worker.py
@@ -27,7 +27,8 @@ from physicalai.export import ExportablePolicyMixin
 from physicalai.train import Trainer
 
 from schemas import Job, Model, Snapshot
-from schemas.job import JobStatus, TrainJobPayload
+from schemas.base_job import JobStatus
+from schemas.job import TrainJobPayload
 from services import DatasetService, ModelService
 from services.event_processor import EventType
 from services.job_service import JobService
